rag/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.36.23
botocore==1.36.23
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.1
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dtreeviz==2.2.2
durationpy==0.9
emoji==2.14.1
et-xmlfile==1.1.0
eval_type_backport==0.2.2
executing==2.0.1
faiss-cpu==1.10.0
fastapi==0.109.2
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.7
ImageHash==4.3.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kubernetes==32.0.0
langchain==0.3.19
langchain-chroma==0.2.2
langchain-community==0.3.17
langchain-core==0.3.36
langchain-google-genai==2.0.9
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.8
llvmlite==0.43.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mlxtend==0.23.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
mypy-extensions==1.0.0
narwhals==1.13.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.12.0
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
posthog==3.14.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.28.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.0.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2023.3.post1
PyWavelets==1.7.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
RapidFuzz==3.12.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.19.0
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.4
streamlit==1.40.1
sympy==1.13.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.49.0
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
unstructured==0.11.6
unstructured-client==0.30.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
virtualenv==20.28.1
visions==0.7.6
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
widgetsnbextension==4.0.11
wordcloud==1.9.4
wrapt==1.17.2
xgboost==2.1.3
xlrd==2.0.1
yarl==1.18.3
ydata-profiling==4.12.1
zipp==3.21.0
zstandard==0.23.0
